requirements.txt
Django~=3.2.2
redis~=3.5.3
pika~=1.2.0
djangorestframework~=3.12.4